Move Google.Logging.Log4Net to Google.Devtools.AspNet.
