+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet.Snippets/CloudLogging/log4net-template.xml
