+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet.Tests/CloudLogging/Log4NetTest.cs
@@ -27,7 +27,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Xunit;
 
-namespace Google.Logging.Log4Net.Tests
+namespace Google.Devtools.AspNet.Tests
 {
     public class Log4NetTest
     {
