+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet/CloudLogging/AsyncAutoResetEvent.cs
@@ -15,7 +15,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Google.Logging.Log4Net
+namespace Google.Devtools.AspNet
 {
     internal sealed class AsyncAutoResetEvent
     {
