+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet/CloudLogging/FileLogQueue.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Google.Logging.Log4Net
+namespace Google.Devtools.AspNet
 {
     // Very TODO
