+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet/CloudLogging/GoogleStackdriverAppender.cs
@@ -24,7 +24,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Google.Logging.Log4Net
+namespace Google.Devtools.AspNet
 {
     /// <summary>
     /// Appends logging events to Google Stackdriver Logging.
