+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet/CloudLogging/MemoryLogQueue.cs
@@ -16,7 +16,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Google.Logging.Log4Net
+namespace Google.Devtools.AspNet
 {
     internal sealed class MemoryLogQueue : ILogQueue
     {
