+++ apis/Google.Devtools.AspNet/Google.Devtools.AspNet/project.json
@@ -22,6 +22,9 @@
     "Google.Api.Gax": "1.0.0-beta04",
     "Google.Devtools.Clouderrorreporting.V1Beta1": { "target": "project" },
     "Google.Devtools.Cloudtrace.V1": "0.3.0",
+    "Google.Logging.Type": "1.0.0-beta04",
+    "Google.Logging.V2": "1.0.0-beta04",
+    "log4net": "2.0.5",
     "Microsoft.AspNet.WebApi.Core": "5.2.3"
   },
