--- apis/Google.Logging.V2/Google.Logging.Log4Net.Snippets/project.json
@@ -1,21 +1,0 @@
-﻿{
-  "description": "Google.Logging.Log4Net.Snippets Class Library",
-
-  "buildOptions": {
-    "keyFile": "../../GoogleApis.snk",
-    "embed": {
-      "include": [ "log4net-template.xml" ]
-    }
-  },
-
-  "dependencies": {
-    "Google.Logging.Log4Net": { "target": "project" },
-    "xunit": "2.2.0-beta2-build3300",
-    "dotnet-test-xunit": "2.2.0-preview2-build1029"
-  },
-  "testRunner": "xunit",
-
-  "frameworks": {
-    "net451": { }
-  }
-}
