--- apis/Google.Logging.V2/Google.Logging.Log4Net.Tests/project.json
@@ -1,20 +1,0 @@
-﻿{
-  "description": "Unit tests for the Google StackDriver log4net appender library.",
-
-  "buildOptions": {
-    "keyFile": "../../GoogleApis.snk"
-  },
-
-  "dependencies": {
-    "Google.Logging.Log4Net": { "target": "project" },
-    "Google.Api.Gax.Testing": "1.0.0-beta04",
-    "xunit": "2.2.0-beta2-build3300",
-    "dotnet-test-xunit": "2.2.0-preview2-build1029",
-    "Moq": "4.6.36-alpha"
-  },
-  "testRunner": "xunit",
-
-  "frameworks": {
-    "net451": { }
-  }
-}
