--- apis/Google.Logging.V2/Google.Logging.Log4Net/LogEntryExtra.cs
@@ -1,30 +1,0 @@
-﻿// Copyright 2016 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Google.Logging.V2;
-
-namespace Google.Logging.Log4Net
-{
-    internal struct LogEntryExtra
-    {
-        public LogEntryExtra(long id, LogEntry entry)
-        {
-            Id = id;
-            Entry = entry;
-        }
-
-        public long Id { get; }
-        public LogEntry Entry { get; }
-    }
-}
