--- apis/Google.Logging.V2/Google.Logging.Log4Net/project.json
@@ -1,29 +1,0 @@
-﻿{
-  "version": "1.0.0-beta04-*",
-  "description": "Log4Net client library for the Google Stackdriver Logging API.",
-  "authors": [ "Google Inc." ],
-
-  "packOptions": {
-    "summary": "Log4Net client library for the Google Stackdriver Logging API.",
-    "owners": [ "google-cloud" ],
-    "tags": [ "Google", "Cloud", "Logging", "log4net", "Stackdriver" ],
-    "iconUrl": "https://cloud.google.com/images/gcp-icon-64x64.png",
-    "projectUrl": "https://github.com/GoogleCloudPlatform/google-cloud-dotnet",
-    "licenseUrl": "http://www.apache.org/licenses/LICENSE-2.0",
-    "requireLicenseAcceptance": false
-  },
-
-  "buildOptions": {
-    "keyFile": "../../GoogleApis.snk",
-    "xmlDoc": true
-  },
-
-  "dependencies": {
-    "Google.Logging.V2": { "target": "project" },
-    "log4net": "2.0.5"
-  },
-
-  "frameworks": {
-    "net45": { }
-  }
-}
